drag and drop (horror) BETA
WARNING flashing lights and any private information shown is not real and randomized WARNING. This is a game about someone in the early 2000's (FAKE STORY) about you getting delivered a package from the company "Windows" but the letter "i" looks a little bit weird, but you don't care. You bring the package in to see a large computer with a keyboard and mouse inside. You decided to plug it in to be greeted with a grey screen with text saying, "the Drag and Drop beta test". You try to exit out but it didn't allow you to, so you decide to continue with the test. The computer tells you to do different tasks with the drag and drop feature as it slowly gets uncanny and creepy.